+++ tfusion/src/cuda/imgproc.cu
@@ -266,7 +266,8 @@ namespace tfusion
                 float yl = (y - c.y) * finv.y;
                 float lambda = sqrtf (xl * xl + yl * yl + 1);
 
-                dists(y, x) = __float2half_rn(depth(y, x) * lambda * 0.001f); //meters
+                //dists(y, x) = __float2half_rn(depth(y, x) * lambda * 0.001f); //meters
+				dists(y, x) = (depth(y, x) * lambda * 0.001f); //meters
             }
         }
     }
